Rockchip: orangepi_5_defconfig: Enable U-Boot EFI loader for *RHEL* AARCH64
Enable EFI_LOADER support.

Whether CONFIG_EFI_LOADER is configured depends on whether CONFIG_X86 or CONFIG_ARM is configured.
When we explicitly configure "#CONFIG_EFI_LOADER is not set" in defconfig, KCONFIG ignores the above
condition. So we chose to remove it from defconfig to enable EFI_LOADER support.

Signed-off-by: gesangtome <gesangtome@foxmail.com>